// api/models/JobCreateSchema.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type JobCreateSchema = {
    company_name: string;
    company_description: string;
    title: string;
    location: string;
    description: string;
    requirements: string;
    salary: string;
    instructions: string;
    deadline: string;
    keywords: string;
    image: string;
    contact_info: string;
    mbti: string;
    job_fields: string;
    holland: string;
    additional: string;
};